Import function update

+++ CresijApp/Scripts/UserInfo.js
@@ -1,5 +1,5 @@
 ﻿$ = jQuery.noConflict();
-$(function GetOrgData() {
+$(document).ready(function GetOrgData() {
     var jsonData = JSON.stringify({
         name: ""
     });
@@ -22,10 +22,10 @@ function OnErrorCall_(respo) {
 }
 
 function FillOrgData(idata) {
-    
+    var userid = sessionStorage.getItem("LoginId");
     var col1 = idata[0], col2 = idata[1], col3 = idata[2],
         col4 = idata[3], col5 = idata[4], col6 = idata[5],
-        col7 = idata[6], col8 = idata[7];
+        col7 = idata[6], col8 = idata[7]; col9= idata[8]
     $("#usertable tr:gt(0)").remove();
     console.log("got user data ");
     //$("#orgTable:not(:first)").remove();
@@ -61,12 +61,20 @@ function FillOrgData(idata) {
         column8.classList = "pl-3";
         column8.innerText = col8[i];
 
+        var linktext = '';
+        if (col2[i] == userid) {
+            linktext = ' <a class="jbtn_xs hover_orange" j-page-box="#jcontent"><i class="fa fa-cog"></i></a> ' +
+                '<a class="jbtn_xs hover_red" j-page-box="#jcontent"><i class="fa fa-trash"></i></a>';
+        }
+        else {
+            linktext = ' <a class="JURL jbtn_xs hover_orange show_jmodal" j-page-href="window/alert/data/userAuthSetting.html?Index=' + col2[i] + '" j-page-box="#jcontent"><i class="fa fa-cog"></i></a>' +
+                '<a class="JURL jbtn_xs hover_red show_jmodal" j-page-href="window/alert/Delete_User.html?Index=' + col2[i] + '" j-page-box="#jcontent"><i class="fa fa-trash"></i></a>';
+        }
         var column9 = row1.insertCell(9);
         column9.classList = "pl-3 hover_btn_td";
         column9.innerHTML = '<a class="JURL jbtn_xs hover_blue show_jmodal" j-page-href="window/alert/data/Edit_UserInfo.html?Index='+col2[i]+'" j-page-box="#jcontent"><i class="fa fa-pencil"></i></a>' +
-            '<a class="JURL jbtn_xs hover_blue show_jmodal" j-page-href="window/alert/data/synchro.html" j-page-box="#jcontent"><i class="fa fa-refresh"></i></a>' +
-            '<a class="JURL jbtn_xs hover_orange show_jmodal" j-page-href="window/alert/data/userAuthSetting.html" j-page-box="#jcontent"><i class="fa fa-cog"></i></a>' +
-            '<a class="JURL jbtn_xs hover_red show_jmodal" j-page-href="window/alert/Delete_User.html?Index=' + col2[i] +'" j-page-box="#jcontent"><i class="fa fa-trash"></i></a>';
+            '<a class="JURL jbtn_xs hover_blue show_jmodal" j-page-href="window/alert/data/synchro.html" j-page-box="#jcontent"><i class="fa fa-refresh"></i></a> ' +
+            linktext;
         
         $("#usertable").find('tbody').append($(row1));
     }
